third_party/blink/web_tests/editing/selection/caret-in-div-containing-empty-block-expected.txt

Testcase for bug 108053: Caret is not displayed when trying to focus inside a contenteditable element containing an empty block. To manually verify the issue, click inside the three boxes. A caret should be displayed for each of them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS horizontalCaretRect.top is textCaretRect.top
PASS horizontalCaretRect.width is textCaretRect.width
PASS horizontalCaretRect.height is textCaretRect.height
PASS verticalCaretRect.top is textCaretRect.left
PASS verticalCaretRect.width is textCaretRect.height
PASS verticalCaretRect.height is textCaretRect.width
PASS successfullyParsed is true

TEST COMPLETE
